refactor(Path Finder): Seperated movement into it's own package as it's too mixed.
