+++ game-api/src/main/kotlin/org/alter/api/ext/PawnExt.kt
@@ -11,6 +11,7 @@ import org.alter.game.model.entity.Npc
 import org.alter.game.model.entity.Pawn
 import org.alter.game.model.entity.Player
 import org.alter.game.model.item.Item
+import org.alter.game.model.move.stopMovement
 import org.alter.game.model.timer.FROZEN_TIMER
 import org.alter.game.model.timer.STUN_TIMER
