+++ game-plugins/src/main/kotlin/org/alter/plugins/content/area/lumbridge/chat/alkharid_guard.plugin.kts
@@ -1,5 +1,7 @@
 package org.alter.plugins.content.area.lumbridge.chat
 
+import org.alter.game.model.move.moveTo
+
 /**
  * Thanks to Tomm for the movement code
  * Thanks to Desetude for gate open and close code
