+++ game-plugins/src/main/kotlin/org/alter/plugins/content/combat/combat.plugin.kts
@@ -1,9 +1,10 @@
 package org.alter.plugins.content.combat
 
-import org.alter.game.action.PawnPathAction
+import org.alter.game.model.move.PawnPathAction
 import org.alter.game.model.attr.COMBAT_TARGET_FOCUS_ATTR
 import org.alter.game.model.attr.FACING_PAWN_ATTR
 import org.alter.game.model.attr.INTERACTING_PLAYER_ATTR
+import org.alter.game.model.move.stopMovement
 import org.alter.game.model.timer.FROZEN_TIMER
 import org.alter.game.model.timer.STUN_TIMER
 import org.alter.plugins.content.combat.specialattack.SpecialAttacks
@@ -55,14 +56,13 @@ suspend fun cycle(it: QueueTask): Boolean {
             }
         }
     }
-
     val strategy = CombatConfigs.getCombatStrategy(pawn)
     val attackRange = strategy.getAttackRange(pawn)
-    val pathFound = PawnPathAction.walkTo(it, pawn, target, interactionRange = attackRange, lineOfSight = false)
+
+    val pathFound = PawnPathAction.walkTo(it, pawn, target, lineOfSightRange = attackRange, lineOfSight = true)
     if (target != pawn.attr[FACING_PAWN_ATTR]?.get()) {
         return false
     }
-
     if (!pathFound) {
         pawn.stopMovement()
         if (pawn.entityType.isNpc) {
@@ -83,11 +83,6 @@ suspend fun cycle(it: QueueTask): Boolean {
         Combat.reset(pawn)
         return false
     }
-
-    pawn.stopMovement()
-
-    // if (pawn is Player && pawn.getEquipment(EquipmentType.WEAPON) != null && world.plugins.executeWeaponCombatLogic(pawn, pawn.getEquipment(EquipmentType.WEAPON)!!.id)) else
-
     if (Combat.isAttackDelayReady(pawn)) {
         if (Combat.canAttack(pawn, target, strategy)) {
             if (pawn is Player && AttackTab.isSpecialEnabled(pawn) && pawn.getEquipment(EquipmentType.WEAPON) != null) {
@@ -98,20 +93,8 @@ suspend fun cycle(it: QueueTask): Boolean {
                 }
                 pawn.message("You don't have enough power left.")
             }
-
-            if (pawn is Player && world.plugins.hasExecItemCmbtLogic(pawn)) {
-                pawn.equipment.rawItems.forEach {
-                    it?.let {
-                        world.plugins.executeItemCombatLogic(pawn, it.id)
-                    }
-                }
-            } else {
-                if (pawn is Player) {
-                    pawn.message("No logic.")
-                }
-                strategy.attack(pawn, target)
-                Combat.postAttack(pawn, target)
-            }
+            strategy.attack(pawn, target)
+            Combat.postAttack(pawn, target)
         } else {
             Combat.reset(pawn)
             return false
