+++ game-plugins/src/main/kotlin/org/alter/plugins/content/commands/commands/admin/tele.plugin.kts
@@ -1,5 +1,6 @@
 package org.alter.plugins.content.commands.commands.admin
 
+import org.alter.game.model.move.moveTo
 import org.alter.game.model.priv.Privilege
 import org.alter.plugins.content.commands.Commands_plugin.Command.tryWithUsage
