+++ game-plugins/src/main/kotlin/org/alter/plugins/content/commands/commands/all/col_grid.plugin.kts
@@ -1,3 +1,4 @@
+import org.alter.game.model.collision.isClipped
 import org.alter.plugins.content.commands.Commands_plugin.Command.tryWithUsage
 
 on_command("col_grid") {
