+++ game-plugins/src/main/kotlin/org/alter/plugins/content/commands/commands/moveNpc.plugin.kts
@@ -1,3 +1,5 @@
+import org.alter.game.model.move.walkTo
+
 val paths =
     listOf(
         Tile(3219, 3222),
