+++ game-plugins/src/main/kotlin/org/alter/plugins/content/objects/Flax.plugin.kts
@@ -1,5 +1,7 @@
 package org.alter.plugins.content.objects
 
+import org.alter.game.model.move.walkTo
+
 /**
  *  @author <a href="https://github.com/CloudS3c">Cl0ud</a>
  *  @author <a href="https://www.rune-server.ee/members/376238-cloudsec/">Cl0ud</a>
