+++ game-plugins/src/main/kotlin/org/alter/plugins/content/skills/firemaking/firemaking.plugin.kts
@@ -2,7 +2,7 @@ package org.alter.plugins.content.skills.firemaking
 
 LogData.values().forEach { log ->
     on_item_on_item(Items.TINDERBOX, log.log) {
-        player.queue { Firemaking.burnLog(this, log) }
+       // player.queue { Firemaking.burnLog(this, log) }
     }
     on_ground_item_option(log.log, "Light") {
         player.graphic(555)
