+++ game-server/src/main/kotlin/org/alter/game/message/handler/OpObjHandler.kt
@@ -1,7 +1,7 @@
 package org.alter.game.message.handler
 
 import net.rsprot.protocol.game.incoming.objs.OpObj
-import org.alter.game.action.GroundItemPathAction
+import org.alter.game.model.move.GroundItemPathAction
 import org.alter.game.message.MessageHandler
 import org.alter.game.model.EntityType
 import org.alter.game.model.Tile
@@ -10,6 +10,7 @@ import org.alter.game.model.attr.INTERACTING_OPT_ATTR
 import org.alter.game.model.entity.Client
 import org.alter.game.model.entity.GroundItem
 import org.alter.game.model.entity.Player
+import org.alter.game.model.move.moveTo
 import org.alter.game.model.priv.Privilege
 import java.lang.ref.WeakReference
