+++ game-server/src/main/kotlin/org/alter/game/message/handler/OpObjTHandler.kt
@@ -1,8 +1,8 @@
 package org.alter.game.message.handler
 
 import net.rsprot.protocol.game.incoming.objs.OpObjT
-import org.alter.game.action.GroundItemPathAction
-import org.alter.game.action.ObjectPathAction
+import org.alter.game.model.move.GroundItemPathAction
+import org.alter.game.model.move.ObjectPathAction
 import org.alter.game.message.MessageHandler
 import org.alter.game.model.EntityType
 import org.alter.game.model.Tile
@@ -11,6 +11,8 @@ import org.alter.game.model.entity.Client
 import org.alter.game.model.entity.GameObject
 import org.alter.game.model.entity.GroundItem
 import org.alter.game.model.entity.Player
+import org.alter.game.model.move.moveTo
+import org.alter.game.model.move.stopMovement
 import org.alter.game.model.priv.Privilege
 import java.lang.ref.WeakReference
