+++ game-server/src/main/kotlin/org/alter/game/model/move/GroundItemPathAction.kt
@@ -1,8 +1,7 @@
-package org.alter.game.action
+package org.alter.game.model.move
 
 import dev.openrune.cache.CacheManager.getItem
 import net.rsprot.protocol.game.outgoing.misc.player.SetMapFlag
-import org.alter.game.model.MovementQueue
 import org.alter.game.model.attr.GROUNDITEM_PICKUP_TRANSACTION
 import org.alter.game.model.attr.INTERACTING_GROUNDITEM_ATTR
 import org.alter.game.model.attr.INTERACTING_ITEM
